[feature] Fix comments threads behavior
