+++ app/src/main/java/com/makentoshe/habrachan/view/post/comments/CommentsFragment.kt
@@ -5,17 +5,14 @@ import android.util.SparseArray
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.core.util.containsKey
 import androidx.core.util.set
-import androidx.core.util.valueIterator
-import androidx.core.view.updateLayoutParams
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
-import com.airbnb.epoxy.*
 import com.makentoshe.habrachan.R
 import com.makentoshe.habrachan.common.entity.comment.Comment
+import com.makentoshe.habrachan.model.post.comment.ArticleCommentEpoxyModel
+import com.makentoshe.habrachan.model.post.comment.ArticleCommentsEpoxyController
 import com.makentoshe.habrachan.ui.post.comments.CommentsFragmentUi
 import com.makentoshe.habrachan.viewmodel.post.comments.CommentsFragmentViewModel
 import io.reactivex.disposables.CompositeDisposable
@@ -33,15 +30,19 @@ class CommentsFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         val recyclerview = view.findViewById<RecyclerView>(R.id.article_comments_recyclerview)
-        val epoxyController = ArticleCommentsEpoxyController(ArticleCommentEpoxyModel.Factory())
+        val epoxyController =
+            ArticleCommentsEpoxyController(
+                ArticleCommentEpoxyModel.Factory()
+            )
+        epoxyController.setContext(requireContext())
 
         viewModel.successObservable.subscribe {
             epoxyController.setComments(parseComments(it.data))
             recyclerview.adapter = epoxyController.adapter
 
-            val maxLevel = it.data.reduce { acc, comment -> if (acc.level < comment.level) comment else acc }.level
-            val step = (recyclerview.width / 2) / maxLevel
-            epoxyController.setStep(step)
+//            val maxLevel = it.data.reduce { acc, comment -> if (acc.level < comment.level) comment else acc }.level
+//            val step = (recyclerview.width / 3) / maxLevel
+//            epoxyController.setStep(step)
 
             epoxyController.requestModelBuild()
         }.let(disposables::add)
@@ -92,89 +93,6 @@ class CommentsFragment : Fragment() {
 
         companion object {
             private const val ID = "Id"
-        }
-    }
-}
-
-class ArticleCommentsEpoxyController(private val factory: ArticleCommentEpoxyModel.Factory) : EpoxyController() {
-
-    private var comments = SparseArray<ArrayList<Comment>>() //= parseComments(commentsResponse.data)
-
-    private var stepSpacePx = 0
-
-    fun setComments(comments: SparseArray<ArrayList<Comment>>) {
-        this.comments = comments
-    }
-
-    fun setStep(stepSpacePx: Int) {
-        this.stepSpacePx = stepSpacePx
-    }
-
-    override fun buildModels() {
-        comments.valueIterator().forEach { comments ->
-            comments.forEach { comment ->
-                factory.build(stepSpacePx, comment).addTo(this)
-            }
-        }
-    }
-
-}
-
-@EpoxyModelClass(layout = R.layout.comments_fragment_comment)
-abstract class ArticleCommentEpoxyModel : EpoxyModelWithHolder<ArticleCommentEpoxyModel.ViewHolder>() {
-
-    @EpoxyAttribute
-    var message = ""
-
-    @EpoxyAttribute
-    var author = ""
-
-    @EpoxyAttribute
-    var score = 0
-
-    @EpoxyAttribute
-    var timePublished = ""
-
-    var stepSpacePx: Int = 0
-
-    override fun bind(holder: ViewHolder) {
-        holder.messageView?.text = message
-        holder.authorView?.text = author
-        holder.scoreView?.text = score.toString()
-        holder.timePublishedView?.text = timePublished
-        holder.rootView?.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-            marginStart = stepSpacePx
-        }
-    }
-
-    class ViewHolder : EpoxyHolder() {
-        var messageView: TextView? = null
-        var authorView: TextView? = null
-        var timePublishedView: TextView? = null
-        var scoreView: TextView? = null
-        var avatarView: ImageView? = null
-        var rootView: View? = null
-
-        override fun bindView(itemView: View) {
-            messageView = itemView.findViewById(R.id.comments_fragment_comment_message)
-            authorView = itemView.findViewById(R.id.comments_fragment_comment_author)
-            timePublishedView = itemView.findViewById(R.id.comments_fragment_comment_timepublished)
-            scoreView = itemView.findViewById(R.id.comments_fragment_comment_score)
-            avatarView = itemView.findViewById(R.id.comments_fragment_comment_avatar)
-            rootView = itemView
-        }
-    }
-
-    class Factory {
-        fun build(stepSpacePx: Int, comment: Comment): ArticleCommentEpoxyModel {
-            val model = ArticleCommentEpoxyModel_()
-            model.id(comment.id)
-            model.message = comment.message
-            model.stepSpacePx = stepSpacePx * comment.level
-            model.author = comment.author.login
-            model.timePublished = comment.timePublished
-            model.score = comment.score
-            return model
         }
     }
 }
