+++ app/src/main/res/layout/comments_fragment_comment.xml
@@ -1,30 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="wrap_content"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:layout_width="match_parent"
-    tools:layout_height="match_parent"
+    android:layout_width="match_parent"
+    android:layout_marginTop="8dp"
     android:layout_height="wrap_content">
+
+    <LinearLayout
+        android:id="@+id/comments_fragment_comment_verticalcontainer"
+        android:layout_width="wrap_content"
+        android:layout_height="0dp"
+        android:orientation="horizontal"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
         android:id="@+id/comments_fragment_comment_avatar"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="@id/comments_fragment_comment_timepublished"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/comments_fragment_comment_author"
-        tools:src="@drawable/ic_account"
         app:layout_constraintDimensionRatio="1:1"
-        />
+        app:layout_constraintStart_toEndOf="@id/comments_fragment_comment_verticalcontainer"
+        app:layout_constraintTop_toTopOf="@id/comments_fragment_comment_author"
+        tools:src="@drawable/ic_account" />
 
     <TextView
         android:id="@+id/comments_fragment_comment_author"
         android:layout_width="0dp"
-        android:maxLines="1"
         android:layout_height="wrap_content"
         android:layout_marginStart="4dp"
+        android:maxLines="1"
         android:textStyle="bold"
+        android:textColor="@android:color/black"
         app:layout_constraintEnd_toStartOf="@+id/comments_fragment_comment_score"
         app:layout_constraintStart_toEndOf="@id/comments_fragment_comment_avatar"
         app:layout_constraintTop_toTopOf="parent"
@@ -50,11 +58,13 @@
 
     <TextView
         android:id="@+id/comments_fragment_comment_message"
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:textColor="@android:color/black"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintStart_toEndOf="@id/comments_fragment_comment_verticalcontainer"
         app:layout_constraintTop_toBottomOf="@+id/comments_fragment_comment_timepublished"
-        tools:text="SAS ASA ANUS PSA" />
+        tools:ignore="WebViewLayout" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
