+++ app/src/main/res/layout/comments_fragment_comment_vertical.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<View android:layout_width="2dp"
+    android:layout_marginEnd="6dp"
+    android:layout_height="match_parent"
+    android:background="@color/colorPrimary"
+    xmlns:android="http://schemas.android.com/apk/res/android" />
